Implementacion de validacion de fecha en formato YYYYMM y tablas (cip01, cp03, consMf, consultExt y censoDmeh)

+++ models/baseCusn11.js
@@ -9,7 +9,7 @@ const BaseCusn11 = sequelize.define('BaseCusn11', {
         primaryKey: true
     },
     cvePresupuestal: {
-        type: DataTypes.INTEGER(12),
+        type: DataTypes.INTEGER(15),
         allowNull: false,
         references: {
             model: 'unidades',
